refactor: useProduct를 생성하여, UI와 비즈니스 로직이 분리 및 파일 이동(Next-Study-habjung#56)
