+++ src/components/profile/container/ProductListContainer.tsx
@@ -1,27 +1,22 @@
 import classNames from 'classnames';
-import { useCallback } from 'react';
 
-import { IProduct, IProductList } from '@/api/types/product';
+import { IProductList } from '@/api/types/product';
 import Text from '@/components/common/text/Text';
 
-import ProductCard from './ProductCard';
+import ProductCard from '../components/product/ProductCard';
+import { useProduct } from '../hook/useProduct';
 
 interface IProductListSectionProps {
   productList: IProductList;
   className?: string;
 }
 
-export default function ProductListSection({
+export default function ProductListContainer({
   productList,
   className,
 }: IProductListSectionProps) {
   const { product } = productList;
-
-  // handleProductClick을 메모이제이션하여 리렌더링이 발생해도 새로운 함수가 생성되는 것을 방지
-  const handleProductClick = useCallback((productData: IProduct) => {
-    console.log(`click: ${productData.itemName}`);
-    // TODO: 상품 클릭시 상품 상세페이지 이동 or 상품 모달
-  }, []);
+  const { productClick } = useProduct();
 
   return (
     <section className={classNames(className)}>
@@ -36,7 +31,7 @@ export default function ProductListSection({
             <ProductCard
               key={productData.id}
               product={productData}
-              onClick={() => handleProductClick(productData)}
+              onClick={() => productClick(productData)}
             />
           ))}
         </div>
